sinusgenyo

+++ src/VizjegView.java
@@ -8,7 +8,7 @@ public class VizjegView extends JatekView{
 	static int y = 250;
 	static int oldal = 100;
 	
-	public void Draw(Graphics g, Integer sorszam) {
+	public void Draw(Graphics g, double sorszam) {
 		/*int x,y,oldal;
 		x=0;
 		y=0;
